refactor: split semantic grammar modules
