+++ src/commands/semantic-grammar/common.ts
@@ -1,0 +1,202 @@
+import type { ElementTarget, InteractionTarget } from '../../client-types.ts';
+import { splitSelectorFromArgs } from '../../daemon/selectors.ts';
+import type { CliFlags } from '../../utils/command-schema.ts';
+import { AppError } from '../../utils/errors.ts';
+import { compactRecord } from '../semantic-common.ts';
+import type {
+  DaemonWriter,
+  SelectionOptions,
+  SemanticDaemonRequest,
+  SemanticRequestInput,
+} from './types.ts';
+
+export function direct(
+  command: string,
+  positionals?: (input: SemanticRequestInput) => string[],
+): DaemonWriter {
+  return (input) => request(command, positionals ? positionals(input) : [], input);
+}
+
+export function request(
+  command: string,
+  positionals: string[],
+  options: SemanticRequestInput,
+): SemanticDaemonRequest {
+  return { command, positionals, options: normalizeCommonRequestOptions(options) };
+}
+
+function normalizeCommonRequestOptions(options: SemanticRequestInput): SemanticRequestInput {
+  return options.deviceTarget !== undefined && options.target === undefined
+    ? { ...options, target: options.deviceTarget }
+    : options;
+}
+
+export function commonInputFromFlags(flags: CliFlags): Record<string, unknown> {
+  return compactRecord({
+    session: flags.session,
+    platform: flags.platform,
+    deviceTarget: flags.target,
+    device: flags.device,
+    udid: flags.udid,
+    serial: flags.serial,
+    iosSimulatorDeviceSet: flags.iosSimulatorDeviceSet,
+    androidDeviceAllowlist: flags.androidDeviceAllowlist,
+  });
+}
+
+export function selectionOptionsFromFlags(flags: CliFlags): SelectionOptions {
+  return {
+    platform: flags.platform,
+    target: flags.target,
+    device: flags.device,
+    udid: flags.udid,
+    serial: flags.serial,
+    iosSimulatorDeviceSet: flags.iosSimulatorDeviceSet,
+    androidDeviceAllowlist: flags.androidDeviceAllowlist,
+  };
+}
+
+export function selectorSnapshotInputFromFlags(flags: CliFlags): Record<string, unknown> {
+  return compactRecord({
+    depth: flags.snapshotDepth,
+    scope: flags.snapshotScope,
+    raw: flags.snapshotRaw,
+  });
+}
+
+export function selectorSnapshotOptionsFromFlags(flags: CliFlags): {
+  depth?: number;
+  scope?: string;
+  raw?: boolean;
+} {
+  return {
+    depth: flags.snapshotDepth,
+    scope: flags.snapshotScope,
+    raw: flags.snapshotRaw,
+  };
+}
+
+export function repeatedInputFromFlags(flags: CliFlags): Record<string, unknown> {
+  return compactRecord({
+    count: flags.count,
+    intervalMs: flags.intervalMs,
+    holdMs: flags.holdMs,
+    jitterPx: flags.jitterPx,
+    doubleTap: flags.doubleTap,
+  });
+}
+
+export function semanticTargetFromClientTarget(
+  target: InteractionTarget | ElementTarget,
+): Record<string, unknown> {
+  if ('ref' in target && target.ref !== undefined) {
+    return compactRecord({ kind: 'ref', ref: target.ref, label: target.label });
+  }
+  if ('selector' in target && target.selector !== undefined) {
+    return { kind: 'selector', selector: target.selector };
+  }
+  const point = target as { x: number; y: number };
+  return { kind: 'point', x: point.x, y: point.y };
+}
+
+export function interactionTargetPositionals(input: InteractionTarget): string[] {
+  if (input.ref !== undefined) return [input.ref, ...optionalString(input.label)];
+  if (input.selector !== undefined) return [input.selector];
+  return [String(input.x), String(input.y)];
+}
+
+export function elementTargetPositionals(input: ElementTarget): string[] {
+  if (input.ref !== undefined) return [input.ref, ...optionalString(input.label)];
+  return [input.selector];
+}
+
+export function readElementTargetFromPositionals(positionals: string[]): ElementTarget {
+  if (positionals[0]?.startsWith('@')) {
+    return { ref: positionals[0], label: optionalTrimmedText(positionals.slice(1)) };
+  }
+  const selector = positionals.join(' ').trim();
+  if (!selector) throw new AppError('INVALID_ARGS', 'get requires @ref or selector expression');
+  return { selector };
+}
+
+export function readGetFormat(value: string | undefined): 'text' | 'attrs' {
+  if (value === 'text' || value === 'attrs') return value;
+  throw new AppError('INVALID_ARGS', 'get only supports text or attrs');
+}
+
+export function splitRequiredSelector(
+  positionals: string[],
+  options: { preferTrailingValue?: boolean } = {},
+) {
+  const split = splitSelectorFromArgs(positionals, options);
+  if (!split) throw new AppError('INVALID_ARGS', 'is requires a selector expression');
+  return split;
+}
+
+export function readJsonObject(value: string, label: string): Record<string, unknown> {
+  try {
+    const parsed = JSON.parse(value) as unknown;
+    if (parsed && typeof parsed === 'object' && !Array.isArray(parsed)) {
+      return parsed as Record<string, unknown>;
+    }
+  } catch {}
+  throw new AppError('INVALID_ARGS', `${label} must be a JSON object`);
+}
+
+export function optionalTrimmedText(values: string[]): string | undefined {
+  const text = values.join(' ').trim();
+  return text || undefined;
+}
+
+export function setOf<T extends string>(...values: T[]): ReadonlySet<T> {
+  return new Set(values);
+}
+
+export function commandNameSet<const TName extends string>(
+  names: readonly TName[],
+): ReadonlySet<string> {
+  return new Set(names);
+}
+
+export function isOneOf<T extends string>(
+  value: string | undefined,
+  values: ReadonlySet<T>,
+): value is T {
+  return value !== undefined && values.has(value as T);
+}
+
+export function isFiniteNumberString(value: string | undefined): boolean {
+  if (value === undefined || value.trim() === '') return false;
+  return Number.isFinite(Number(value));
+}
+
+export function readFiniteNumber(value: string | undefined, label: string): number | undefined {
+  if (value === undefined) return undefined;
+  const parsed = Number(value);
+  if (Number.isFinite(parsed)) return parsed;
+  throw new AppError('INVALID_ARGS', `${label} must be a finite number.`);
+}
+
+export function optionalCliNumber(value: string | undefined): number | undefined {
+  return value === undefined ? undefined : Number(value);
+}
+
+export function optionalString(value: string | undefined): string[] {
+  return value === undefined ? [] : [value];
+}
+
+export function optionalNumber(value: number | undefined): string[] {
+  return value === undefined ? [] : [String(value)];
+}
+
+export function requiredString(value: string | undefined, message: string): string {
+  if (value === undefined || value === '') throw new AppError('INVALID_ARGS', message);
+  return value;
+}
+
+export function requiredDaemonString(value: unknown, message: string): string {
+  if (typeof value !== 'string' || value.length === 0) {
+    throw new AppError('INVALID_ARGS', message);
+  }
+  return value;
+}
